change parameter type of countclient class

+++ src/main/java/com/usa/ciclo3/retociclo3/reports/CountClient.java
@@ -5,19 +5,19 @@ import com.usa.ciclo3.retociclo3.model.Client;
 
 
 public class CountClient {
-    private Double total;
+    private Long total;
     private Client client;
 
-    public CountClient(Double total, Client client) {
+    public CountClient(Long total, Client client) {
         this.total = total;
         this.client = client;
     }
 
-    public Double getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
